FinalstausをEndGameに統合、EndGameをPatchに移動
